Auth is now User.  Cosmetic?  Maybe it's slightly better, it allows me to dump user attributes with auth concerns.
I'm really not having fun with Cognito but I think it's just a matter of inexperience.  I want to get past authorization, then I want to clean it up.  For whatever reason I thought I should make a User instead of an Auth and then do the exact same thing inside of the User.  The idea is that the User will represent more aspects of the client than just whether or not they're authorized.  I need to get more in the habit of just jamming stuff in and refactoring after instead of when I've barely started.