put Grid in headers, removed it from everywhere else
This way I don't have to put Grid throughout the app.  I kept the Welcome page outside of the header path because I want the splash to go across the entire screen.  Originally I put Grid into App but it caused headers to not go across the entire screen (obviously).